requirements.txt
accelerate>=0.16.0
torchvision
transformers>=4.25.1
ftfy
tensorboard
Jinja2
peft==0.7.0
diffusers==0.33.1
bitsandbytes
wandb
